Create week-04-response.md
In this image, the three messages conveyed are linguistic, denoted, and connoted messages.  The linguistic message is that the meal is originating from America, more than likely the south, given that's where soul food originated. For the denoted message, I would say this would be dinner, since an entrée, and multiple sides are shown on a plate, implying that it's more than likely dinner time, when a hearty meal such as this is served.  Lastly, the connoted message would be the fact that this meal is associated with black culture. Given that this meal was made famous by, and prepared by African Americans, this makes it the connoted message. Lastly, this message would be read differently if there were higher definition present, as this could potentially show things such as grease, moisture, texture, etc. within the picture on the food, to give us a better idea of it's characteristics.